+++ src/main/java/ajou/mse/dimensionguard/exception/player/PlayerNotFoundByMemberIdException.java
@@ -3,7 +3,7 @@ package ajou.mse.dimensionguard.exception.player;
 import ajou.mse.dimensionguard.exception.common.NotFoundException;
 
 public class PlayerNotFoundByMemberIdException extends NotFoundException {
-    public PlayerNotFoundByMemberIdException(Integer memberId) {
+    public PlayerNotFoundByMemberIdException(Long memberId) {
         super("memberId=" + memberId);
     }
 }
