+++ src/main/java/ajou/mse/dimensionguard/exception/room/AlreadyParticipatingException.java
@@ -4,7 +4,7 @@ import ajou.mse.dimensionguard.exception.common.ConflictException;
 
 public class AlreadyParticipatingException extends ConflictException {
 
-    public AlreadyParticipatingException(Integer participatingRoomId) {
+    public AlreadyParticipatingException(Long participatingRoomId) {
         super("참여중인 방의 id=" + participatingRoomId);
     }
 }
