+++ src/main/java/ajou/mse/dimensionguard/exception/room/RoomIdNotFoundException.java
@@ -3,7 +3,7 @@ package ajou.mse.dimensionguard.exception.room;
 import ajou.mse.dimensionguard.exception.common.NotFoundException;
 
 public class RoomIdNotFoundException extends NotFoundException {
-    public RoomIdNotFoundException(Integer roomId) {
+    public RoomIdNotFoundException(Long roomId) {
         super("roomId=" + roomId);
     }
 }
