+++ src/main/java/ajou/mse/dimensionguard/repository/redis/InGameRequestStatusRepository.java
@@ -3,5 +3,5 @@ package ajou.mse.dimensionguard.repository.redis;
 import ajou.mse.dimensionguard.dto.redis.InGameRequestStatus;
 import org.springframework.data.repository.CrudRepository;
 
-public interface InGameRequestStatusRepository extends CrudRepository<InGameRequestStatus, Integer> {
+public interface InGameRequestStatusRepository extends CrudRepository<InGameRequestStatus, Long> {
 }
